bug 1529342: remove tests and fix docstrings for things that will be changed
I've got a PR that's too big. I'm taking out some parts from that PR and
landing them on their own. In the future PR, I'm removing the
PolyStorage crashstorage, MetricsCounter, the benchmarking wrapper, and
cleaning up the task manager code.

This removes the tests for those things and fixes the doc strings in the
task manager code.

This doesn't affect the run-time behavior of webapp, processor, or
crontabber. There are no functionality/behavior changes in this PR.