bug 1529342: update webapp/crashstats/crashstats/ tests
This gets pytest running so it's not dying on import errors. Then this
continues and updates all the webapp/crashstats/crashstats/ tests.

While doing this, I removed the bulk of the mocking code and broke out
the report view tests into its own class (not a unittest class!). This
resulted in me fixing a bunch of buggy test data and improved the tests
and the usefulness of the tests dramatically.

I also adjusted the API for the test helpers.