bug 1529342: update webapp/crashstats/topcrashers/tests
This rewrites the topcrashers tests. I took out all the mocking. I also
broke up a monolithic test that touched a lot of different things into
separate tests that were more comprehensive.

For example, the startup crash icons are important to get right, so this
rewrote the tests for that to cover the various possibilities.