bug 1529342: adjust crash destination settings
This adjsuts the crash destination settings so that the crash storage
items are top-level in the settings file and then referenced from there.
This reduces the long multi-part strings in the tests to access class
options. It also makes mozilla_settings a little more flexible for later
on when we will (probably) need to switch between AWS and GCP things
with a feature flag.