bug 1529342: handle empty string not being a valid int
The webapp imports mozilla_settings and when it does that, it doesn't
have settings for the processor-related configuration, so those things
get set to the empty string which isn't a valid int.

This accounts for that by handling the empty string case and making it a
None. Then it adds some code to the relevant parts of the codebase that
use those settings to handle the None value.