bug 1529342: switch from ACCESS to OPEN; minor improvements
This moves debugging logging to an `is_verbose` conditional block which
reduces some work in the hot loop.

More importantly, this switches from using ACCESS to using OPEN to
update the access time in the LRU. There's one OPEN but multiple
ACCESSes when using a file, so this will radically reduce the number of
inotify events being handled.

This also wraps another os operation in a FileNotFoundError check.