bug 1529342: make removing watches more resilient; set default to 40gb
This makes removing watches a bit more resilient and adds them to the
relevant places. If I understand inotify correctly, when an inode is
removed, any watches for it are also removed automatically. But
something is going on, so I figured I'd add calls to remove_watch in
case I'm wrong about that.

This also sets the default to 40gb. This is a stop gap until I add
parser to Everett for data sizes and we move the cache size to infra.

This also tweaks the logging output to make some things easier to
understand in the logs.