bug 1529342: fix possible race condition when adding directories
We keep running out of space in a way that suggests the cache manager
somehow misses adding a watch for a top-level directory.

The theory is that if multiple stackwalkers are adding subdirectories,
the cache manager would handle a CREATE | ISDIR event, but because
inventory_existing does a directory traversal which takes a while to
run, it's possible that an additional directory gets added, but missed
by the traversal. That missed directory would never get watched and it
will accrue data that isn't accounted for in the LRU cache until it eats
up all the disk.

This changes the handling for CREATE | ISDIR to stop using
inventory_existing and instead create a series of fake inotify Events
just for the contents of the directory created. The fake events will get
handled by the loop and cover adding watches for new directories and
handling new files. The smaller focus should reduce the likelihood of a
race condition causing a missed watch.