bug 1816751: adjust structure of get_version_context_for_product
This adjusts the way it figure out featured versions in relation to the
product setting. Now it iterates over the list and expands "auto" and
(soon) "product_details" in place. This is a little easier to reason
about especially now that we're going to have multiple things that can
expand.