bug 1827416: fix support for cpu_microcode_version
minidump-stackwalk v0.16.0 changed the type of cpu_microcode_version in
the output from a u32 to a hexstring. It also now handles falling back
to the CPUMicrocodeVersion, so we don't need to do that in the processor
rule anymore.

Because we have existing data from the v0.15.0 version of the
stackwalker, we need to include integer as a possible type in the
schema. We can remove that in November 2023 when all the data has
expired.