bug 1837409: fix create-a-bug links for Fenix Java crashes
The JavaStackTrace annotation is a long vaguely structured string. The
create-a-bug code uses that to build the bug description. If that value
is too long, it exceeds the max url length and then gets clipped causing
the create-a-bug links to not work.

This truncates the JavaStackTrace value.

At some point, we should look at switching to JavaException which is
structured when it's available and fall back to JavaStackTrace if
necessary.