bug 1837423: rename IntegerField and improve field tests
This renames IntegerField to NumberField so the name matches the query
type. This makes it a little clearer when skimming code.

This removes the mapping for the flag query type--that no longer exists.

This improves the tests breaking test up into field groups and covering
the __null__ and !__null__ possibilities for BooleanField.