bug 1837423: fix __null__ for number fields
This fixes __null__ and !__null__ for number fields. Support was
partially implemented a couple of years ago in the search library code,
but the Django form fields which parse and validate the request query
string were never adjusted to support __null__ and !__null__, so it
never worked.

This adds support for __null__ and !__null__ and also adds appropriate
validation for combinations of these values and other filters for a
number field. This adds appropriate tests as well.

This also fixes the case where a duplicate filter raised a
ValidationError. If you pass in something like <10 <10, that's not
great, but it shouldn't be a validation error of 10 < 10 since it's not
a non-overlapping range.