bug 1838302: fix cache keys for nested protected data
The protected crash contains several nested structures and the path
through those structures is reflected in the "field" value. Because of
that, the cache_key could be arbitrarily long which memcached doesn't
like.

This adjusts the code to validate the dataset is one of two values and
also hashes the field so it meets memcache key constraints.