bug 1667997: change report_type default to "any"
This changes the report_type default from "crash" which is what we
eventually want to use to "any" which is less terrifying when there's no
data in the search index with a report_type set.

Once we've accumulated several weeks of crash reports with report_type
set, we can change the default back to "crash".