bug-1849266: change generate_create_bug_url to truncate symbols and refactor
This changes how frame symbols in a stack are rendered in the bug
comment from dropping function arguments (which was buggy and didn't
work right in various cases) to truncating at 80 characters.

This also refactors how generate_create_bug_url works. It takes out the
jinja template from generating the comment and some of the hacks we had
to alleviate issues with that.

It reimplements that in a libbugzilla library. It also radically cleans
up the arguments being passed around so now generate_create_bug_url only
requires a crash_report_url, a template, and a processed crash. That's a
lot less nebulous than what we had before and makes it much easier for
us to extend and for others to extract for external uses.