bug 1850428: rework stackwalker execution and timeout
This reworks how we run the stackwalker. Previously, it was run using
the timeout gnu coreutils command which would kill the stackwalker
process after some timeout if it was still running.

processor -> timeout -> stackwalker

Now it runs the stackwalker using the timeout argument to
subprocess.run() and subprocess.run() keeps track of time using threads.
This removes an intermediary process and a possibility for the
stackwalker process to get orphaned.

processor -> stackwalker