bug 1847429: implement signature generation for JavaException
This fixes signature generation to degrade to using the JavaException
data if there is no JavaStackTrace. It maintains the same format we use
for Java signatures.

Currently, it only looks at the first frame (the thing that raised the
exception) in the first exception value (the exception at the beginning
of the cascade), but the scaffolding is there to extend it to iterate
over the frames like we do with the CSignatureTool. We'll use that when
we rework Java signature generation.