bug 1853662: update stackwalker to 0.18.0 (v20230927.0)
This updates the stackwalker to 0.18.0 and also adds a line to the
set_up_stackwalker.sh script so we can see what version it is when we
build the Docker image. Plus (bonus!) if running:

minidump-stackwalk --version

fails because of missing glibc, we'll know at build rather than run
time.