bug 1855494: add psql make target and site status admin page
The psql make target makes it easier to access the db in a local dev
environment. It's something we had in Tecken, but not in Socorro.

I added a section for settings which isn't wildly useful, but does make
it easier to see if DEBUG=True which would be bad in a server
environment.

I added a section for table counts to list tables in the db and their
row counts. This will help us see table counts over time and know when
things are awry.