bug 1858717: switch stackwalker to use output file
This switches the way Socorro runs the stackwalker to use an output file
for output (previously used stdout) and a log file for error output
(previously used stderr). This simplifies the code plus it makes it
easier to differentiate between the output artifact and stackwalker
logging.

This fixes the problem with bin/run_mdsw.sh where it didn't create the
symbols cache directories before running the stackwalker. Now it does.

This also changes bin/run_mdsw.sh to use --verbose=debug since we're
likely using it to debug stackwalker issues.