bug 1863956: remove product support content, clean up docs
Initially, I wanted to move the "add a new product/remove a product"
documentation to Confluence. So I did that. Then in doing that, I ended
up restructuring the documentation home page, merging in the "Sign up"
information, adding a link to the footer, and changing "Documentation"
to "Help" to start making it more support-focused.