bug-1862204: add missing_symbols field
This adds a missing_symbols field to the processed crash and the search
index.

This field is a `;` delimited list of "filename/version/debugid" tuples.

This is generated from the json_dump.modules list of modules that have a
"filename" and "missing_symbols=True". "missing_symbols=True" indicates
that the stackwalker was unable to find a symbols file at the time it
was processing the minidump.

Because I'm working with debug ids, I decided to add a debug_id_helper
pytest fixture for creating valid debug ids.