bug-1873658: fix params parsing date=>=2024-01-10 bug
This fixes a bug that occurs when breaking up the querystring into
key=val parameters when the value has an = in it. Instead of dropping
everything after the second = from the value, this takes more care to
split on the first = and ensures everything after that is part of the
value.