bug-1812771: fix os_name, os_version, and os_pretty_version for Android
We want to pull the version for Android from the `Android_Version`
annotation which holds the sdk version and the release name. Then we use
that to build an os_pretty_version that works for C++/Rust crash reports
as well as Java crash reports and is much more useful than what we've
got now.