bug-1809726: update Django to 4.2
This updates Django to 4.2. In order to do that, we also had to update
django-pipeline to 3.0.0 so it worked correctly. But then that caused
problems when building static files. I managed to fix the jssha error,
but couldn't fix the jquery.jsonview one.

jquery.jsonview is unmaintained and hasn't had a version release in 8
years and the example urls redirect to terrible places. I switched Crash
Stats to use jquery.json-viewer which required some minor js and css
changes, but generally works the same.