bug 1883055: stop using django-session-csrf
We used django-session-csrf which is a Mozilla maintained library that
we wrote back in the heady Mozilla webdev days which had properties our
webdev and websec folks wanted. That was a long time ago, Django's
built-in csrf module has improved over time, the world has changed (no
more Flash), and the library hasn't been actively maintained in some
time.

This drops the django-session-csrf and some of the interesting code
around it in favor of Django's built-in csrf.