bug-1100790: fix supersearch for default operator and platform field
The platform field is analyzed for completely unclear reasons even
though it only has one of five values: Windows NT, Android, Linux, Mac
OS X, or Unknown.

If we do a supersearch with the default operator on "Linux", it works
fine--it returns crash reports where os_name="Linux"

If we do a supersearch with the default operator on "Mac OS X", it also
works fine--it returns crash reports where os_name="Mac OS X".

If we do a supersearch with the default oprator on "Linux" and "Mac OS
X", it only returns Linux results because there are two values and so it
switches to a terms filter on an analyzed field which will never match
"Mac OS X".

This fixes that by changing the default operator to use "match" queries
in all cases (single value and multi-value) because "match" queries will
analyze the search text with the same analyzer used for the search
field.

This caused one test to fail, but I contend that the test was buggy, so
I fixed it.

I also added a comment to the "platform" super search field definition
about how the platform field shouldn't be analyzed. We can figure that
out another day.