bug 1894925: add handling for TimeoutError
This adds handling for TimeoutError so it'll go through the loop again.
It increases the blocking timeout to 2 which is slightly longer than 1s.
It also changes everything to f-strings.

The theory is that in CI, it's taking longer than 1s for localstack to
respond sometimes which causes CI to fail and because the loop didn't
handle TimeoutError, it would die immediately.