bug-1889185: implement stage submitter and fake collector
This implements a stage submitter that pulls crash ids from QUEUE and
crash data from STORAGE, packages that up into a crash report payload,
and submits it to STAGE_SUBMITTER_DESTINATIONS.

This creates a `make runsubmitter` system that runs the stage submitter
and a fake collector allowing you to use `bin/process_crash.sh` to
set everything up and publish a crash id to the standard queue which the
submitter will see and submit a crash to the fake collector. The fake
collector will parse the payload and log headers, annotations, and
files.

This adds stage submitter tests.