bug-1898341, bug-1898345: add "host" tag, fix metrics key prefix
This adds a `host` tag to emitted metrics when running in GCP. This is
derived from `HOSTNAME` if it exists, otherwise it defaults to
`socket.gethostname()` like our other services.

This changes Sentry and logging to use `HOSTNAME` configuration variable
rather than `HOST_ID`. This brings us in line with other services as we
migrate to GCP.

This also adds `"socorro"` prefix to all emitted keys, but only for the
GCP environments. This brings keys in line with our other services.

In order to do this, I had to create a singleton `METRICS` and then
rework everything to use that.