bug-1877291: process crash reports we discover weren't processed
verifyprocessed is a Django management command that runs every day,
looks at the previous day of raw crashes, checks to see if there's a
processed crash and data in Elasticsearch, and adds a
MissingProcessedCrash record if there wasn't.

This fixes the command so it also tosses the crash id in the
reprocessing queue.