bug-1900646: fix submitter to only consume standard queue
This fixes the submitter to only consume from the standard subscription
rather than from standard, priority, and reprocessing.

It does it by stomping on the subscription_name settings for priority
and reprocessing in the PubSubCrashQueue class. That means this will
never run in AWS, but that's what our plan was anyhow.