bug-1900995: remove SignatureParentIDNotEqualsChildID rule
This removes a signature generation rule that used moz_crash_reason data
to stomp on the signature in a certain situation. The assertion that
triggered crash reports for this situation is no longer in the code base
so this rule no longer has any effect. This removes the rule and the use
of moz_crash_reason.