bug-1100790: improve supersearch operator docs and forms
This changes "has terms" and "does not have terms" in the super search
form to "matches" and "does not match" to match the change in code.
Generally, it's a lot clearer as to what it's doing, too.

This also changes "is" and "is not" in the super search form to "is
exactly" and "is not exactly" which more closely matches what the code
is doing.

This also improves the clarify of the super search operator docs so it's
easier to understand what super search operators do.