bug-1901998: fix buginfo view to stop sending sentry errors (#6779)
When Bugzilla is in a mood, the buginfo view queries the BugInfo API and
that API throws a requests.exceptions.RetryError which the buginfo view
didn't handle and thus we end up with a Sentry error.

When this happens, the report view Bugzilla tab gets back an HTTP 500
from querying the buginfo view and kind of shrugs and doesn't show bug
details. The user isn't really impacted--they can continue their merry
day.

Bugzilla seems to be down a lot, so we get these Sentry errors we're not
going to do anything with. Thus I did a couple of things:

1. increase the backoff_factor for Bugzilla requests giving us a little
   more time for Bugzilla to find its happy place
2. change the buginfo view to handle the RetryError by returning an HTTP
   500 with an error message and _not_ sending a Sentry event