bug-1901997: handle invalid BuildID for redacted crash reports (#6777)
* Add support for --pause to process_crashes.sh

This makes it possible to adjust a crash report for testing a specific
processing scenario.

* bug-1901997: handle invalid BuildID for redacted crash reports

If the user viewing the crash report does not have protected data access
and the crash report has a crash annotation value that doesn't match our
schema, then the redaction process can fail. In this scenario, Crash
Stats can't recover and redact the document properly.

In bug 1901997, we were seeing this scenario when BuildID=null. Because
crash annotation data comes from the crash reporter which might be
running on a machine that has bad hardware, this scenario can happen
with other fields.

These changes handle that scenario and shows a page indicating the crash
report is malformed and providing a link the user can use to report the
malformed crash report.