bug-1928808: fix _report_type validation (#6784)
This fixes _report_type validation in the TopCrashers form so it doesn't
create bad Super Search queries which kick up an HTTP 500.

Previously, the form did no useful validation (why even have a form?).
Now it enforces that _report_type is one of "any", "hang", or "crash"
which are the only three values it should have. If it gets a different
value, the form will cause the view to raise an HTTP 400 with an error
saying that the _report_type value isn't one of the available choices.