bug-1928934: fix super search fields in signature report forms (#6786)
* bug-1928934: fix super search fields in signature report forms

In 9b425e50, I changed the name of the field the webapp permissions were
stored in, but missed updating the signature report view. This fixes
that and adds a test.

* bug-1928934: fix get_fields() return value shape

This returns `fields` to being a list of field names so the rest of the
code that depends on it works correctly.