bug-1906021: strip whitespace from Android_CPU_ABI2 and other string fields (#6824)
This changes the CopyFromRawCrashRule to strip the right side of string
field values. I haven't seen anything in crash reports so far where
whitespace at the end of a crash annotation value is significant so I
think this is safe.

If it turns out not to be safe, we can extend CopyFromRawCrashRule to
look at a "transforms" field to glean how it should transform the value
after copying it. This would be helpful for the increasing number of
fields that are comma or semi-colon delimited, too.