bug-1933774: omit collector-added fields from Crash Annotations tab (#6823)
* bug-1933774: omit collector-added fields from Crash Annotations tab

This removes the collector-added fields from the Crash Annotations tab
of the report view. It's confusing that these were there.

This also adds "version" to the Debug tab. This value is the version of
the raw crash structure in storage. It's relevant to Socorro internals,
but nothing else.

* Remove old code

This said to remove it after 2022. It's after 2022 now.