Fix TestCustomQuery test that fails at beginning of year (#6853)
* Fix TestCustomQuery test that fails at beginning of year

The index name template uses week number and when we calculate
the expected index names by adding 7 days to figure out the next week
number, that doesn't work at the beginning of the year when the previous
year includes a partial week.

This fixes that by changing the code to figure out week numbers by
incrementing by one day, generating index names, throwing those in a
set, and then turning that into a sorted list.

* Fix range so index names extend across entire date range