obs-443: switch to docker compose watch
Currently for the local dev environment, we volume mount the local
directory as `/app` which causes anything built during docker image
build to be shadowed by the user's local directory. We get around this
by having docker image build any artifacts into directories in the image
that are not under `/app`. This is a little wonky and there are hacks
strewn about to alleviate the consequences.

Instead of doing that, this PR is an initial pass and switching to
docker compose watch. In this new model, we no longer volume mount the
local directory as `/app` and instead docker compose watches for changes
on the local file system and syncs them with the container and will run
any rebuilding scripts.

This should work with the local dev environment, CI, and server
environments.

This requires v2.22.0 of docker compose plugin.

This doesn't update documentation and doesn't document edge cases to be
aware of.