bug-1945653: adjust processing queue balance
This changes the way the processor pulls from queues. Previously, it
pulled pull_max_messages from each queue in order and then processed
that. If we had a backlog of standard and reprocessing queue work, then
they'd be handled equally. That's bad for standard queue processing and
caused alerts and stuff.

This changes it to prefer processing the standard queue 5 over the
reprocessing queue 5 to 1 and makes it much less likely the standard
queue backs up when someone reprocesses a bunch of crash reports.