bug-1954596: update stackwalker to v0.25.0
This updates the stackwalker to v0.25.0. One thing this picks up is a
new field in the stackwalker output: soft_errors.

However, the value for that field is documented as "object" and we don't
know what that is so we can't put it in the schema. If they want that
field to be public or if they want us to do something with it in Crash
Stats, they'll have to flesh out the structure of the object.